frontend-api/pyproject.toml
[tool.poetry]
name = "frontend-api"
version = "0.1.0"
description = ""
authors = ["Precious Ohanyere <obbyprecious24@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.114.0"}
psycopg2-binary = "^2.9.9"
pyjwt = "^2.9.0"
sqlalchemy = "^2.0.34"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
